fix(controllers/sauce.js) - rewrite deleteSauce + ajout error message si id sauce inexistant

+++ backend/controllers/sauce.js
@@ -22,25 +22,47 @@ exports.createSauce = (req, res, next) => {
         .catch(error => { res.status(400).json({ error }) })
 };
 
-// DELETE Suppression d'une sauce
-exports.deleteSauce = (req, res, next) => {
-    Sauce.findOne({ _id: req.params.id })
-        .then(sauce => {
+// // DELETE Suppression d'une sauce
+// exports.deleteSauce = (req, res, next) => {
+//     Sauce.findOne({ _id: req.params.id })
+//         .then(sauce => {
+//             if (sauce.userId != req.auth.userId) {
+//                 res.status(401).json({ message: 'Non autorisé' });
+//             } else {
+//                 const filename = sauce.imageUrl.split('/images/')[1];
+//                 fs.unlink(`images/${filename}`, () => {
+//                     Sauce.deleteOne({ _id: req.params.id })
+//                         .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
+//                         .catch(error => res.status(401).json({ error }));
+//                 });
+//             }
+//         })
+//         .catch(error => {
+//             res.status(500).json({ error });
+//         });
+// };
+
+exports.deleteSauce = async (req, res, next) => {
+    try {
+        const sauce = await Sauce.findOne({ _id: req.params.id })
+        if (sauce === null) {
+            res.status(401).json({ message: `La sauce n'existe pas` });
+        } else {
             if (sauce.userId != req.auth.userId) {
-                res.status(401).json({ message: 'Not authorized' });
+                res.status(401).json({ message: 'Non autorisé' });
             } else {
                 const filename = sauce.imageUrl.split('/images/')[1];
                 fs.unlink(`images/${filename}`, () => {
                     Sauce.deleteOne({ _id: req.params.id })
                         .then(() => { res.status(200).json({ message: 'Objet supprimé !' }) })
                         .catch(error => res.status(401).json({ error }));
-                });
+                })
             }
-        })
-        .catch(error => {
-            res.status(500).json({ error });
-        });
-};
+        };
+    } catch (error) {
+        res.status(500).json({ error });
+    }
+}
 
 // POST Like sauce
 exports.likeSauce = (req, res, next) => {
